Revert "Merged PR 3649: Reworked this old, wide, table-based topic into a long page with subsections, one for each first-level option
For [12651840](https://microsoft.visualstudio.com/OS/_workitems?id=12651840), I reworked this old, wide, table-based topic into a long page with subsections, one for each first-level option. Due to Markdown and our build system, this presentation results in a page that is readable in narrow page widths, unlike the rendering of the previous version. This version has a simple table that lists and briefly describes the first-level options, and each of those option names is linked to the subsection with more detail.

I also corrected various issues, made some repeated language more consistent, and added descriptions for two additional options that were not included in the previous documentation. I changed the presentation of syntax and placeholders to match style guidelines, changed capitalized text to lowercase when the usage was caseless (which is all of it), and added links to each subsection that takes the browser back to the table of options at the top of the file.

This pull request applies the version of powercfg-command-line-options.md that I worked on in branch v-gmoor-1709 before abandoning that branch (due to not being able to resolved a conflict). This change addresses bug 12651840."