Add OpenSSH (Beta) to list of v2 FODs
Tried my best to follow the formatting of the file. Gave as much detail as provided by NetFx3 (which is to say, very little), but if you want to add a link to our project (where we have a wiki, more info, etc.), that's fine. 

At some point we need to come back here when our OpenSSH port exits beta to remove the beta tag.