This change updates the Cortana and Bluetooth feature information
Starting with RS5 version of Windows IoT Core, the Cortana application is no longer available as an option.  This change marks both as deprecated and recommends the user consult with the Cortana Dev Center.  

Starting in RS5 there are multiple, individual, Bluetooth features available for Windows IoT.  This change includes the new feature identifiers